core: DoctrineEntityPersister refactoring
